Desarrollo de pagina ingreso de tickets masivos en base de datos

+++ src/script/bd.sql
@@ -119,7 +119,7 @@ CREATE TABLE compra (
   hin_rut int(10) NOT NULL,
   usr_username varchar(20) NOT NULL,  
   com_monto int(6) NOT NULL,
-  com_token int NOT NULL,
+  com_token bigint NOT NULL,
   com_tipo varchar(1) NOT NULL,
   com_anulada varchar(1) NOT NULL,
   com_fecha datetime NOT NULL,
